Update spatiotemporal simulator exports
Move the promoted CNS, GPE, and Gray-Scott simulators into the
stable spatiotemporal namespace. Update first-party Hydra targets, API
docs, and tests now that these simulators are canonical. Keep
experimental import paths as compatibility wrappers during the
transition.
