--- docs/reference/experimental/gray_scott.rst
@@ -1,7 +1,0 @@
-autosim.experimental.simulations.gray_scott
-===========================================
-
-.. automodule:: autosim.experimental.simulations.gray_scott
-   :members:
-   :undoc-members:
-   :show-inheritance:
