--- docs/reference/experimental/gross_pitaevskii.rst
@@ -1,7 +1,0 @@
-autosim.experimental.simulations.gross_pitaevskii
-=================================================
-
-.. automodule:: autosim.experimental.simulations.gross_pitaevskii
-   :members:
-   :undoc-members:
-   :show-inheritance:
