--- docs/reference/experimental/navier_stokes_conditioned.rst
@@ -1,7 +1,0 @@
-autosim.experimental.simulations.navier_stokes_conditioned
-==========================================================
-
-.. automodule:: autosim.experimental.simulations.navier_stokes_conditioned
-   :members:
-   :undoc-members:
-   :show-inheritance:
