+++ docs/reference/simulations/spatiotemporal/gray_scott.rst
@@ -1,0 +1,7 @@
+autosim.simulations.spatiotemporal.gray_scott
+======================================================================
+
+.. automodule:: autosim.simulations.spatiotemporal.gray_scott
+   :members:
+   :undoc-members:
+   :show-inheritance:
