+++ docs/reference/simulations/spatiotemporal/gross_pitaevskii.rst
@@ -1,0 +1,7 @@
+autosim.simulations.spatiotemporal.gross_pitaevskii
+======================================================================
+
+.. automodule:: autosim.simulations.spatiotemporal.gross_pitaevskii
+   :members:
+   :undoc-members:
+   :show-inheritance:
