+++ docs/reference/simulations/spatiotemporal/index.md
@@ -1,0 +1,3 @@
+# Spatiotemporal Simulations
+
+Stable simulator implementations that return spatiotemporal data.
