+++ docs/reference/simulations/spatiotemporal/navier_stokes_conditioned.rst
@@ -1,0 +1,7 @@
+autosim.simulations.spatiotemporal.navier_stokes_conditioned
+======================================================================
+
+.. automodule:: autosim.simulations.spatiotemporal.navier_stokes_conditioned
+   :members:
+   :undoc-members:
+   :show-inheritance:
