+++ src/autosim/configs/simulator/gpe/exploratory/high_complexity.yaml
@@ -2,7 +2,7 @@
 # Gaussian wavepacket, not the physical ground state. This does not correspond
 # to a realistic experimental setup. Useful as a stress-test / complexity sweep
 # but not as a faithful BEC experiment.
-_target_: autosim.experimental.simulations.GrossPitaevskiiEquation2D
+_target_: autosim.simulations.spatiotemporal.GrossPitaevskiiEquation2D
 return_timeseries: true
 log_level: warning
 output_names: [density, real, imag]
