+++ src/autosim/configs/simulator/gpe/exploratory/rotating_box_instability.yaml
@@ -18,7 +18,7 @@
 #     Phys. Rev. Lett. 92, 060401 (2004). DOI: 10.1103/PhysRevLett.92.060401
 #   Lundh, E., et al. "Dynamic instability of a rotating Bose-Einstein condensate."
 #     Phys. Rev. A 64, 033612 (2001). DOI: 10.1103/PhysRevA.64.033612
-_target_: autosim.experimental.simulations.GrossPitaevskiiEquation2D
+_target_: autosim.simulations.spatiotemporal.GrossPitaevskiiEquation2D
 return_timeseries: true
 log_level: warning
 output_names: [density, real, imag]
