+++ src/autosim/configs/simulator/gpe/exploratory/speckle_only.yaml
@@ -1,6 +1,6 @@
 # Exploratory speckle-only regime:
 # stationary disorder drives spatially inhomogeneous dynamics without obstacle forcing.
-_target_: autosim.experimental.simulations.GrossPitaevskiiEquation2D
+_target_: autosim.simulations.spatiotemporal.GrossPitaevskiiEquation2D
 return_timeseries: true
 log_level: warning
 output_names: [density, real, imag]
