+++ src/autosim/configs/simulator/gpe/exploratory/vortex_lattice.yaml
@@ -5,7 +5,7 @@
 # Physical Review Letters 84, 806 (2000). DOI: 10.1103/PhysRevLett.84.806
 # Abo-Shaeer, J. R., et al. "Observation of vortex lattices in Bose-Einstein condensates." 
 # Science 292, 476-479 (2001). DOI: 10.1126/science.1060182
-_target_: autosim.experimental.simulations.GrossPitaevskiiEquation2D
+_target_: autosim.simulations.spatiotemporal.GrossPitaevskiiEquation2D
 return_timeseries: true
 log_level: warning
 output_names: [density, real, imag]
