+++ src/autosim/configs/simulator/gpe/laser_only_wake.yaml
@@ -46,7 +46,7 @@
 #     DOI: 10.1088/1367-2630/ac8656
 #
 # =============================================================================
-_target_: autosim.experimental.simulations.GrossPitaevskiiEquation2D
+_target_: autosim.simulations.spatiotemporal.GrossPitaevskiiEquation2D
 return_timeseries: true
 log_level: warning
 output_names: [density, real, imag]
