+++ src/autosim/configs/simulator/gpe/rotating_box_lattice.yaml
@@ -23,7 +23,7 @@
 #   J. Phys.: Condens. Matter 31 275401 (2019). DOI: 10.1088/1361-648X/ab14c5
 #   Simula, T. et al. "Giant vortex lattice deformations in rapidly rotating Bose-Einstein condensates."
 #   Phys. Rev. Lett. 92, 060401 (2004). DOI: 10.1103/PhysRevLett.92.060401
-_target_: autosim.experimental.simulations.GrossPitaevskiiEquation2D
+_target_: autosim.simulations.spatiotemporal.GrossPitaevskiiEquation2D
 return_timeseries: true
 log_level: warning
 output_names: [density, real, imag]
