+++ src/autosim/simulations/spatiotemporal/__init__.py
@@ -1,0 +1,11 @@
+"""Spatiotemporal simulator implementations."""
+
+from .gray_scott import GrayScott
+from .gross_pitaevskii import GrossPitaevskiiEquation2D
+from .navier_stokes_conditioned import ConditionedNavierStokes2D
+
+__all__ = [
+    "ConditionedNavierStokes2D",
+    "GrayScott",
+    "GrossPitaevskiiEquation2D",
+]
